# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BoardGameTools API
  description: |
    REST API fuer BoardGameTools — Brettspiel-Verwaltung, Sessions, Events mit Voting, Gruppen und Spielereihen.

    ## Authentifizierung
    Die meisten Endpunkte erfordern eine gueltige NextAuth.js Session (Cookie-basiert).
    Oeffentliche Endpunkte unter `/api/public/` und `/api/auth/` sind ohne Login erreichbar.

    ## Fehlerformat
    Alle Fehler folgen dem Format: `{ "error": "Deutsche Fehlermeldung" }`

    ## Rate Limiting
    Oeffentliche Endpunkte sind per IP rate-limited (Upstash Redis + In-Memory Fallback).
  version: 0.40.1
  contact:
    name: BoardGameTools
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Lokale Entwicklung
  - url: https://boardgametools.vercel.app
    description: Production
...
components:
...
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js Session Cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: API Access Token aus `POST /api/mobile/v1/auth/login` oder `refresh`
...
security:
  - session: []
paths:
...
  /api/mobile/v1/auth/apple:
    post:
      tags: [Mobile Auth]
      summary: Sign in with Apple
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                identityToken: { type: string }
                authorizationCode: { type: string }
                name: { type: string }
                deviceName: { type: string }
      responses:
        '200': { description: Token-Paar und Nutzerprofil }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /api/mobile/v1/events/{id}/proposals:
    get:
      tags: [Mobile]
      summary: Spielvorschlaege
      security: [ bearerAuth: [] ]
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      responses:
        '200': { description: Vorschlaege }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Mobile]
      summary: Spielvorschlag hinzufuegen
      security: [ bearerAuth: [] ]
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gameId: { type: string }
                bggId: { type: string }
                bggName: { type: string }
      responses:
        '201': { description: Vorschlag erstellt }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/mobile/v1/events/{id}/votes:
    post:
      tags: [Mobile]
      summary: Fuer Spielvorschlag abstimmen
      security: [ bearerAuth: [] ]
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                proposalId: { type: string }
      responses:
        '200': { description: Abstimmung gespeichert }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/mobile/v1/groups/{id}/join:
    post:
      tags: [Mobile]
      summary: Gruppe per Share-Token beitreten
      security: [ bearerAuth: [] ]
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shareToken: { type: string }
      responses:
        '200': { description: Beigetreten }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/mobile/v1/groups/{id}/comments:
    get:
      tags: [Mobile]
      summary: Gruppen-Kommentare
      security: [ bearerAuth: [] ]
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      responses:
        '200': { description: Kommentare }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Mobile]
      summary: Kommentar hinzufuegen
      security: [ bearerAuth: [] ]
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content: { type: string }
                pollId: { type: string }
      responses:
        '201': { description: Kommentar erstellt }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/mobile/v1/bgg/search:
    get:
      tags: [Mobile BGG]
      summary: BGG-Suche
      security: [ bearerAuth: [] ]
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: Suchergebnisse }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/mobile/v1/bgg/lookup:
    post:
      tags: [Mobile BGG]
      summary: EAN/UPC via BGG aufloesen
      security: [ bearerAuth: [] ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ean: { type: string }
      responses:
        '200': { description: Gefundenes BGG-Spiel }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/mobile/v1/bgg/import:
    post:
      tags: [Mobile BGG]
      summary: Spiel aus BGG importieren
      security: [ bearerAuth: [] ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bggId: { type: string }
      responses:
        '201': { description: Importiertes Spiel }
        '401': { $ref: '#/components/responses/Unauthorized' }
...